requirements.txt
streamlit
# Mic recorder / browser STT component (free, client-side)
streamlit-mic-recorder
# Google Gemini client (new interface)
google-generativeai
requests
